requirements.txt
flask
flask-cors
tensorflow-cpu
opencv-python-headless
pillow
numpy
gunicorn
werkzeug